Lots of changes, lots of functionality and javascript
Working on adding information to the html page, I added images, and a footer, I also changed how I was doing logos that go on the same line as text, I was using css to make a span have a background image that was the logo, but instead I changed it to an image tag that has display: inline on it. I also managed to add code that would change the link to maroon on hover no matter what color it is using filter css. I made the slideshow functional and works both automatically and manually, it constantly changes its image but when you hover over it the image gets bigger and the slideshow stops, something I always want slideshows to do but they don't. Lastly I added things I add to every site I make, a horizontal scroll bar on the top of the page and a to top button. I added lots of javascript for most of the things I mentioned.